server/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic>=2.9.0
python-jose==3.3.0
redis==5.0.3
structlog==24.2.0
prometheus-client==0.20.0
websockets==12.0
u-msgpack-python==2.8.0
SQLAlchemy>=2.0
asyncmy>=0.2 ; sys_platform != 'win32'
aiomysql>=0.2
alembic>=1.13
passlib[bcrypt]==1.7.4


